data-modeling/dm-postgres/metrics.py
import time
from functools import wraps
from memory_profiler import memory_usage
import log
logger = log.get_logger(__name__,"metrics.log")

def profile(fn):
    @wraps(fn)
    def inner(*args, **kwargs):
        fn_kwargs_str = ', '.join(f'{k}={v}' for k, v in kwargs.items())
        logger.info(f'\n{fn.__name__}({fn_kwargs_str})')

        # Measure time
        t = time.perf_counter()
        retval = fn(*args, **kwargs)
        elapsed = time.perf_counter() - t
        logger.info(f'Time   {elapsed:0.4}')

        # Measure memory
        mem, retval = memory_usage((fn, args, kwargs), retval=True, timeout=200, interval=1e-7)

        logger.info(f'Memory {max(mem) - min(mem)}')
        return retval

    return inner