requirements.txt
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
cassandra-driver==3.22.0
click==7.1.1
configparser==4.0.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ez-setup==0.9
geomet==0.1.2
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
ndjson==0.3.1
notebook==6.0.3
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
utils==1.0.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
